requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.7
pymilvus==2.3.4
python-dotenv==1.0.0
httpx==0.25.2
python-multipart==0.0.6
marshmallow==3.20.1
setuptools